extract SaleSearchCardForm
